combine predict controller, add check GPU, clean docker file in plugin
