--- easyearth_plugin/docker-compose.yml
@@ -1,22 +1,0 @@
-version: '3.8'   # More recent features
-
-services:
-  easyearth-server:
-    image: easyearth_plugin_easyearth-server
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - "3781:3781"
-    volumes:
-      - .:/usr/src/app # Sync the application code
-      - ${DATA_DIR:-./data}:/usr/src/app/data # Data directory
-      - ${TEMP_DIR:-./tmp}:/usr/src/app/tmp  # Temporary directory
-      - ${LOG_DIR:-./logs}:/usr/src/app/logs  # Log directory
-      - ${MODEL_DIR:-~/.cache/easyearth/models}:/usr/src/app/.cache/models  # Model directory
-    environment:
-      - PYTHONUNBUFFERED=1
-      - EASYEARTH_DATA_DIR=/usr/src/app/data
-      - EASYEARTH_TEMP_DIR=/usr/src/app/tmp
-      - EASYEARTH_LOG_DIR=/usr/src/app/logs
-      - MODEL_CACHE_DIR=/usr/src/app/.cache/models
