+++ easyearth_plugin/plugin.py
@@ -13,6 +13,7 @@ from qgis.core import (QgsVectorLayer, QgsGeometry,
 from qgis.gui import QgsMapToolEmitPoint, QgsRubberBand
 from .core.utils import setup_logger
 from .core.prompt_editor import BoxMapTool
+from .core.model_manager import ModelManager
 import json
 import logging
 import os
@@ -20,6 +21,7 @@ import requests
 import shutil
 import subprocess
 import time
+import torch
 
 class EasyEarthPlugin:
     def __init__(self, iface):
@@ -44,8 +46,8 @@ class EasyEarthPlugin:
         self.project_name = "easyearth_plugin"
         self.sudo_password = None  # Add this to store password temporarily
         self.docker_path = 'docker' if shutil.which('docker') else '/Applications/Docker.app/Contents/Resources/bin/docker' # adds compatibility for macOS
-        # self.docker_hub_image_name = "maverickmiaow/easyearth"
-        self.docker_hub_image_name = "lgordon99/easyearth"
+        self.docker_hub_image_name = "maverickmiaow/easyearth"
+        # self.docker_hub_image_name = "lgordon99/easyearth"
         self.docker_mode = True
 
         # Initialize map tools and data
@@ -93,6 +95,7 @@ class EasyEarthPlugin:
         self.predictions_geojson = None
         self.predictions_layer = None
         self.prompts_geojson = None
+        self.prompts_layer = None
 
         # initialize image crs and extent...
         self.raster_crs = None
@@ -181,7 +184,6 @@ class EasyEarthPlugin:
             service_layout.addLayout(status_layout)
             self.docker_run_btn = QPushButton("Start docker")
             self.docker_run_btn.clicked.connect(self.run_or_stop_container)
-            self.docker_run_btn.hide()
             service_layout.addWidget(self.docker_run_btn)
 
             # API Information
@@ -189,8 +191,7 @@ class EasyEarthPlugin:
             api_label = QLabel("API Endpoints:")
             api_label.setStyleSheet("font-weight: bold;")
             self.api_info = QLabel(f"Base URL: http://0.0.0.0:{self.server_port}/v1/easyearth\n"
-                                   f"Infer with point or box prompts: /sam-predict\n"
-                                   f"Infer with no prompts: /segment-predict\n"
+                                   f"Inference: /predict\n"
                                    f"Health check: /ping")
             self.api_info.setWordWrap(True)
             api_layout.addWidget(api_label)
@@ -230,11 +231,11 @@ class EasyEarthPlugin:
             # Image source selection
             source_layout = QHBoxLayout()
             source_label = QLabel("Source:")
-            self.source_dropdown = QComboBox()
-            self.source_dropdown.addItems(["File", "Layer", "Link"])
-            self.source_dropdown.currentTextChanged.connect(self.on_image_source_changed)
+            self.source_combo = QComboBox()
+            self.source_combo.addItems(["File", "Layer", "Link"])
+            self.source_combo.currentTextChanged.connect(self.on_image_source_changed)
             source_layout.addWidget(source_label)
-            source_layout.addWidget(self.source_dropdown)
+            source_layout.addWidget(self.source_combo)
             image_layout.addLayout(source_layout)
 
             # File input
@@ -385,15 +386,17 @@ class EasyEarthPlugin:
             # Update the layer (raster) list in the combo box whenever a layer is added or removed
             QgsProject.instance().layersAdded.connect(self.update_layer_dropdown)
             QgsProject.instance().layersRemoved.connect(self.update_layer_dropdown)
-            # QgsProject.instance().layersAdded.connect(self.auto_zoom_to_selected_layer)
 
             # Connect to QGIS quit signal
             QgsApplication.instance().aboutToQuit.connect(self.cleanup_docker)
 
-            # Start periodic server status check
-            self.status_timer = QTimer()
-            self.status_timer.timeout.connect(self.check_server_status)
-            self.status_timer.start(5000)  # Check every 5 seconds
+            # # Check if the container is running on startup
+            # self.inspect_running_container()
+
+            # Check GPU availability on startup
+            self.check_gpu_availability()
+
+            self.logger.debug("Finished initGui setup")
         except Exception as e:
             self.logger.error(f"Error in initGui: {str(e)}")
             self.logger.exception("Full traceback:")
@@ -404,6 +407,18 @@ class EasyEarthPlugin:
             self.dock_widget.hide()
         else:
             self.dock_widget.show()
+
+    def check_gpu_availability(self):
+        """Check if GPU is available for model inference"""
+        if torch.backends.mps.is_available():
+            self.iface.messageBar().pushMessage("Info", "MPS GPU is available", level=Qgis.Info, duration=5)
+            return "mps"
+        elif torch.cuda.is_available():
+            self.iface.messageBar().pushMessage("Info", "CUDA GPU is available", level=Qgis.Info, duration=5)
+            return "cuda"
+        else:
+            self.iface.messageBar().pushMessage("Warning", "No GPU available, using CPU", level=Qgis.Warning, duration=5)
+            return "cpu"
 
     def check_server_status(self):
         """Check if the server is running by pinging it"""
@@ -432,19 +447,19 @@ class EasyEarthPlugin:
             self.server_running = False
             self.docker_running = False
             self.docker_run_btn.setText("Start docker")
-        
-        self.docker_run_btn.show()
 
     def run_mode_selected(self, mode):
         """Handle run mode selection (Docker or Local)"""
         self.server_group.show()  # Show server group when a mode is selected
+        # Start periodic server status check
+        self.status_timer = QTimer()
+        self.status_timer.timeout.connect(self.check_server_status)
+        self.status_timer.start(5000)  # Check every 5 seconds
 
         if mode == 'docker':
             self.local_mode_button.setChecked(False)
             self.docker_mode = True
-
-            if self.server_status.text() != 'Checking...':
-                self.docker_run_btn.show()
+            self.docker_run_btn.show()
         elif mode == 'local':
             self.docker_mode_button.setChecked(False)
             self.docker_mode = False
@@ -524,8 +539,7 @@ class EasyEarthPlugin:
                 self.browse_button.show()
                 self.layer_dropdown.hide()
                 self.download_button.hide()
-                self.image_path.clear()
-                self.image_path.setPlaceholderText("No image selected")
+                self.initialize_image_path()
                 self.initialize_embedding_path()
                 self.deactivate_embedding_section() if not self.is_sam_model() else None
             elif text == "Layer":
@@ -541,7 +555,7 @@ class EasyEarthPlugin:
                 self.browse_button.hide()
                 self.download_button.show()
                 self.layer_dropdown.hide()
-                self.image_path.setPlaceholderText("Enter image URL (http/https...)")
+                self.image_path.setPlaceholderText("Enter image URL (http/https)...")
                 self.image_path.clear()
                 self.initialize_embedding_path()
                 self.deactivate_embedding_section() if not self.is_sam_model() else None
@@ -588,7 +602,7 @@ class EasyEarthPlugin:
             self.downloading_progress_status.hide()
 
             # Switch to File mode and update path
-            self.source_dropdown.setCurrentText("File")
+            self.source_combo.setCurrentText("File")
             self.image_path.setText(save_path)
             self.load_image()
         except Exception as e:
@@ -704,11 +718,14 @@ class EasyEarthPlugin:
                 )
                 self.logger.info(f"Will save new embedding to: {embedding_path}")
             else:
+                # Check if embedding file exists
                 if os.path.exists(embedding_path):
                     # Found existing embedding
                     self.load_embedding_radio.setChecked(True)
                     self.embedding_path_edit.setEnabled(True)
                     self.embedding_browse_btn.setEnabled(True)
+
+                    # Set the embedding path in the text box
                     self.embedding_path_edit.setText(embedding_path)
 
                     self.iface.messageBar().pushMessage(
@@ -721,7 +738,7 @@ class EasyEarthPlugin:
                 else:
                     # No existing embedding
                     self.no_embedding_radio.setChecked(True)
-                    self.load_embedding_radio.setEnabled(False)
+                    self.load_embedding_radio.setEnabled(False)  # TODO: what if the user have the embeddings saved somewhere else?
                     # clear the embedding path
                     self.embedding_path_edit.clear()
                     self.embedding_browse_btn.setEnabled(False)
@@ -739,38 +756,18 @@ class EasyEarthPlugin:
             self.logger.error(f"Error updating embeddings: {str(e)}")
             QMessageBox.critical(None, "Error", f"Failed to update embeddings: {str(e)}")
 
-    def on_image_selected(self):
-        selected_layer = QgsProject.instance().mapLayersByName(os.path.basename(self.image_path.text()))[0]
-        self.iface.setActiveLayer(selected_layer) # set the selected layer as the active layer
-        self.iface.messageBar().pushMessage(f"Selected layer {selected_layer.name()} with ID {selected_layer.id()}", level=Qgis.Info)
-        
-        if selected_layer.crs() != self.project_crs:
-            selected_layer.setCrs(self.project_crs)  # Set the layer CRS to match the project CRS
-        
-        extent = selected_layer.extent()
+    def initialize_image_path(self):
+        """Reinitialize the image path input field"""
 
-        if not extent.isEmpty():
-            # Add small buffer around the layer (5%)
-            width = extent.width()
-            height = extent.height()
-            
-            if width > 0 and height > 0:
-                buffer_x = width * 0.05
-                buffer_y = height * 0.05
-                
-                extent.setXMinimum(extent.xMinimum() - buffer_x)
-                extent.setXMaximum(extent.xMaximum() + buffer_x)
-                extent.setYMinimum(extent.yMinimum() - buffer_y)
-                extent.setYMaximum(extent.yMaximum() + buffer_y)
-            
-            # Set the canvas extent
-            self.iface.mapCanvas().setExtent(extent)
-            self.iface.mapCanvas().refresh()
+        self.image_path.clear()
+        self.image_path.setPlaceholderText("No image selected")
+        self.image_path.setEnabled(True)
+        self.browse_button.setEnabled(True)
 
     def browse_image(self):
         """Open file dialog for image selection"""
         try:
-            initial_dir = self.data_dir
+            initial_dir = self.data_dir if os.path.exists(self.data_dir) else ""
             file_path, _ = QFileDialog.getOpenFileName(
                 self.dock_widget,
                 "Select Image File",
@@ -800,16 +797,14 @@ class EasyEarthPlugin:
             if not image_path:
                 return
 
-            raster_layer = QgsRasterLayer(image_path, os.path.basename(image_path)) # loads the image as a raster layer
+            # Load the image as a raster layer
+            raster_layer = QgsRasterLayer(image_path, os.path.basename(image_path))
 
             if not raster_layer.isValid():
                 QMessageBox.warning(None, "Error", "Invalid raster layer")
                 return
 
-            if len(QgsProject.instance().mapLayersByName(raster_layer.name())) == 0: # checks if the layer already exists
-                QgsProject.instance().addMapLayer(raster_layer) # adds raster layer to the project
-            
-            self.on_image_selected()
+            QgsProject.instance().addMapLayer(raster_layer) # adds raster layer to the project
 
             # Get image crs and extent
             self.raster_extent, self.raster_width, self.raster_height, self.raster_crs = self.get_current_raster_info(raster_layer)
@@ -967,7 +962,7 @@ class EasyEarthPlugin:
                 # Show coordinates in message bar
                 self.iface.messageBar().pushMessage("Point Info",
                                                     f"Map coordinates: ({point.x():.2f}, {point.y():.2f})\n"
-                                                    f"Pixel coordinates: ({px}, {py})",
+                                                    f"Pixel coordinates sent to server: ({px}, {py})",
                                                     level=Qgis.Info,
                                                     duration=3)
 
@@ -1011,8 +1006,8 @@ class EasyEarthPlugin:
         self.prompt_count = self.prompt_count - 1 if self.prompt_count > 0 else 0  # decrements the prompt counter
         self.add_features_to_layer([], "prompts")
 
-        if len(self.predictions_geojson.get('features')) > 0:
-            self.predictions_geojson['features'] = self.predictions_geojson['features'][0] # removes the last point from the predictions geojson
+        if self.realtime_checkbox.isChecked():
+            self.predictions_geojson['features'] = self.predictions_geojson['features'][:-1] # removes the last point from the prompts geojson
             self.add_features_to_layer([], "predictions")
 
     def on_box_drawn(self, box_geom, start_point, end_point):
@@ -1064,7 +1059,7 @@ class EasyEarthPlugin:
         # Show message bar with box coordinates
         self.iface.messageBar().pushMessage(
             "Box Info",
-            f"Box coordinates: ({pixel_x}, {pixel_y}, {pixel_x + pixel_width}, {pixel_y + pixel_height})",
+            f"Box coordinates sent to server: ({pixel_x}, {pixel_y}, {pixel_x + pixel_width}, {pixel_y + pixel_height})",
             level=Qgis.Info,
             duration=3
         )
@@ -1105,45 +1100,6 @@ class EasyEarthPlugin:
             self.get_prediction(prompt)
 
         return box_geom
-
-    def on_layer_selected(self, index):
-        """Handle layer selection change and check for existing embeddings"""
-        try:
-            if index > 0: # 0th index is "Select a layer..."
-                layer_id = self.layer_dropdown.itemData(index)
-                selected_layer = QgsProject.instance().mapLayer(layer_id) if layer_id else None
-
-                if not selected_layer:
-                    return
-
-                # Get image crs and extent
-                self.raster_extent, self.raster_width, self.raster_height, self.raster_crs = self.get_current_raster_info(selected_layer)
-                msg = (
-                    f"Extent: X min: {self.raster_extent.xMinimum()}, X max: {self.raster_extent.xMaximum()}, "
-                    f"Y min: {self.raster_extent.yMinimum()}, Y max: {self.raster_extent.yMaximum()}; "
-                    f"Width: {self.raster_width}; "
-                    f"Height: {self.raster_height}; "
-                    f"CRS: {self.raster_crs.authid()}"
-                )
-                self.iface.messageBar().pushMessage(
-                    "Selected raster CRS and dimensions",
-                    msg,
-                    level=Qgis.Info,
-                    duration=5
-                )
-
-                self.image_path.setText(selected_layer.source())
-                self.create_prediction_layers()
-
-                # Check for existing embedding
-                layer_source = selected_layer.source()
-                image_name = os.path.splitext(os.path.basename(layer_source))[0]
-                self.update_embeddings(image_name)
-                self.on_image_selected()
-        except Exception as e:
-            self.logger.error(f"Error handling layer selection: {str(e)}")
-            self.logger.exception("Full traceback:")
-            QMessageBox.critical(None, "Error", f"Failed to handle layer selection: {str(e)}")
 
     def add_features_to_layer(self, features, layer_type='prompts', crs=None):
         """
@@ -1292,7 +1248,6 @@ class EasyEarthPlugin:
 
                 if not layer.isValid():
                     raise ValueError("Failed to create valid vector layer")
-                
                 QgsProject.instance().addMapLayer(layer)
                 setattr(self, layer_attr, layer)
             else:
@@ -1373,7 +1328,12 @@ class EasyEarthPlugin:
             # if there are boxes and points in the prompts, we need to run the prediction for both
             if self.is_sam_model():
                 if len(prompts) == 0:
-                    self.iface.messageBar().pushMessage("Info", "No prompts found. Please draw points or boxes.", level=Qgis.Info, duration=3)
+                    self.iface.messageBar().pushMessage(
+                        "Info",
+                        "No prompts found. Please draw points or boxes.",
+                        level=Qgis.Info,
+                        duration=3
+                    )
                     return
                 else:
                     # Check if there are both points and boxes
@@ -1418,6 +1378,8 @@ class EasyEarthPlugin:
             # Show loading indicator
             self.iface.messageBar().pushMessage("Info", "Getting prediction...", level=Qgis.Info)
             QApplication.setOverrideCursor(Qt.WaitCursor)
+
+            # Get the image path and convert for container
             image_path = self.image_path.text()
 
             if not os.path.exists(image_path):
@@ -1467,6 +1429,7 @@ class EasyEarthPlugin:
 
             if self.model_path:
                 payload["model_path"] = self.model_path
+                payload["model_type"] = "sam" if self.is_sam_model() else "segment"
 
             # Show payload in message bar
             if prompts is None or len(prompts) == 0:
@@ -1491,12 +1454,7 @@ class EasyEarthPlugin:
 
             # Send request to SAM server
             try:
-                if prompts is None or len(prompts) == 0:
-                    # No prompts, run full image prediction
-                    predict_url = f"{self.server_url}/segment-predict"
-                else:
-                    # With prompts, run prompt-based prediction
-                    predict_url = f"{self.server_url}/sam-predict"
+                predict_url = f"{self.server_url}/predict"
 
                 response = requests.post(predict_url, json=payload, timeout=6000000)
 
@@ -1677,6 +1635,46 @@ class EasyEarthPlugin:
             self.logger.exception("Full traceback:")
             QMessageBox.critical(None, "Error", f"Failed to prepare layers: {str(e)}")
 
+    def on_layer_selected(self, index):
+        """Handle layer selection change and check for existing embeddings"""
+        try:
+            if index > 0: # 0th index is "Select a layer..."
+                layer_id = self.layer_dropdown.itemData(index)
+                selected_layer = QgsProject.instance().mapLayer(layer_id) if layer_id else None
+
+                if not selected_layer:
+                    return
+
+                # Get image crs and extent
+                self.raster_extent, self.raster_width, self.raster_height, self.raster_crs = self.get_current_raster_info(selected_layer)
+                msg = (
+                    f"Extent: X min: {self.raster_extent.xMinimum()}, X max: {self.raster_extent.xMaximum()}, "
+                    f"Y min: {self.raster_extent.yMinimum()}, Y max: {self.raster_extent.yMaximum()}; "
+                    f"Width: {self.raster_width}; "
+                    f"Height: {self.raster_height}; "
+                    f"CRS: {self.raster_crs.authid()}"
+                )
+                self.iface.messageBar().pushMessage(
+                    "Selected raster CRS and dimensions",
+                    msg,
+                    level=Qgis.Info,
+                    duration=5
+                )
+
+                self.image_path.setText(selected_layer.source())
+                self.create_prediction_layers()
+
+                # Check for existing embedding
+                layer_source = selected_layer.source()
+
+                image_name = os.path.splitext(os.path.basename(layer_source))[0]
+                self.update_embeddings(image_name)
+
+        except Exception as e:
+            self.logger.error(f"Error handling layer selection: {str(e)}")
+            self.logger.exception("Full traceback:")
+            QMessageBox.critical(None, "Error", f"Failed to handle layer selection: {str(e)}")
+
     def cleanup_docker(self):
         """Clean up Docker resources when unloading plugin"""
         try:
@@ -1765,7 +1763,7 @@ class EasyEarthPlugin:
                 QgsProject.instance().removeMapLayer(self.prompts_layer.id())
             if self.predictions_layer:
                 QgsProject.instance().removeMapLayer(self.predictions_layer.id())
-            
+
             # Remove temporary files
             for file_path in [self.temp_prompts_geojson, self.temp_predictions_geojson]:
                 if file_path and os.path.exists(file_path):
